fix: SymbolicExpressionAir findings
- force exact symbolic-row presence for variable-free expression trees by constraining proof-shape AirId lookup counts to VK-derived values
- reject is_present = 1 on dummy cached rows by requiring valid cached flags
- move NodeKind into symbolic_expression/air.rs so the canonical discriminants live with the audited AIR logic
- remove the unused claim_bus and eq_neg_internal_bus wiring
